🔨 Refactor grapher_checks function to etl.grapher.helpers and update validation logic
